+++ src/main/resources/db/migration/V1_6_0__new_table_permissions.sql
@@ -32,3 +32,5 @@ INSERT INTO permissions(per_id, entity, action, lib_id) VALUES
 (7, 'COMPONENT', 'CREATE', NULL),
 (8, 'COMPONENT_TEMPLATE', 'CREATE', NULL),
 (9, 'LIBRARY', 'ACCESS', NULL);
+
+ALTER SEQUENCE permissions_per_id_seq RESTART WITH 10;